Split out genesis assets module
